+++ android/app/src/main/kotlin/deckers/thibault/aves/HomeWidgetConfigureActivity.kt
@@ -3,12 +3,17 @@ package deckers.thibault.aves
 import android.appwidget.AppWidgetManager
 import android.content.Intent
 import android.os.Bundle
+import deckers.thibault.aves.utils.FlutterUtils
+import deckers.thibault.aves.utils.FlutterUtils.enableSoftwareRendering
 import io.flutter.plugin.common.MethodChannel
 
 class HomeWidgetSettingsActivity : MainActivity() {
     private var appWidgetId = AppWidgetManager.INVALID_APPWIDGET_ID
 
     public override fun onCreate(savedInstanceState: Bundle?) {
+        if (FlutterUtils.isSoftwareRenderingRequired()) {
+            intent.enableSoftwareRendering()
+        }
         super.onCreate(savedInstanceState)
 
         // cancel if user does not complete widget setup